+++ example/lib/gen/voyager_gen.dart
@@ -3,24 +3,22 @@ import 'package:flutter/widgets.dart';
 import 'package:provider/provider.dart';
 import 'package:voyager/voyager.dart';
 
-class VoyagerPaths {
-  static const String pathHome = "/home";
-  static const String typeHome = "home";
-  static String pathOther(String title) {
-    return "/other/$title";
-  }
-
-  static const String typeOther = "other";
-  static const String pathFab = "/fab";
-  static const String typeFab = "fab";
-  static const String pathTalks = "/talks";
-  static const String typeTalks = "talks";
-  static String pathObjectItem(String className) {
-    return "/_object/$className";
-  }
-
-  static const String typeObjectItem = "object_item";
+const String pathHome = "/home";
+const String typeHome = "home";
+String pathOther(String title) {
+  return "/other/$title";
 }
+
+const String typeOther = "other";
+const String pathFab = "/fab";
+const String typeFab = "fab";
+const String pathTalks = "/talks";
+const String typeTalks = "talks";
+String pathObjectItem(String className) {
+  return "/_object/$className";
+}
+
+const String typeObjectItem = "object_item";
 
 class VoyagerData extends Voyager {
   VoyagerData({String path, Voyager parent, Map<String, dynamic> config})